% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Returning a data frame table when reading in a FARS csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A filename must be given to carry out this 'fars_read' function. This can be a string and/or numeric.}
}
\value{
A table of the class 'data frame'
}
\description{
The 'fars_read' function returns a table of the class 'data frame' when the input is a filename from the Fatality Analysis Reporting System
(FARS) of the US National Highway Traffic Safety.
}
\details{
When the file name exists, the fars_read function will be applied havng the following steps.
First, the file will be evaluated whether it is a csv file with comma as separator by the 'suppressMessages' function from the base package.
If the file is a csv, the file will be read in using the 'read_csv' function from the readr package and assigned as 'data'.
Then, this 'data' will be converted in a table having a 'data frame' class, using the 'tbl_df' function from the dplyr package.
}
\note{
stop When the file name does not exist. The non-existence of the file name will be reported by
the sentence "file <filename> does not exist".
}
\examples{
fars_read(farsReadTest_example("accident_2013.csv.bz2"))

}

