% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Returning a list of valid years and their months in a FARS file}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{- The 'years' parameter is a vector (atomic or a list) or an expression object}
}
\value{
A list of valid years and their months

#@example
#fars_read_years(c(2013, 2014))
}
\description{
The 'fars_read_years' function returns a list of each element of the 'year' variable which appears in the read-in file.
}
\details{
Within the 'fars_read_years' function, a 'lapply' function from the base package is used. The parameters for
the 'lapply' function are 'years' and the 'make_filename' function (defined above).
Within the 'lapply' function, the 'make_filename' function will be first applied on the 'year' variable and assigned
this as 'file'. Then, the 'tryCatch' function is called and evaluates its content, containing an expression and the function(e)
whic implies an error. The expression is composed of:
     1) the 'fars_read' function of a file as 'dat'
     2) data manipulation of the 'dat' file: adding a new variable 'year' using the 'mutate' function of the dplyr package,
                                            followed by selecting the variables 'MONTH' and 'year' from the 'dat' file
When the year is valid, a list of valid years and months will be returned.
}
\note{
error - When the year is invalid, a warning will be given such as "invalid year: <year>".  Nothing (=NULL) will be returned.
}

