% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExample.R
\name{farsReadTest_example}
\alias{farsReadTest_example}
\title{Path to farsReadTest examples}
\usage{
farsReadTest_example(filename)
}
\arguments{
\item{filename}{Name of file}
}
\description{
Pakage 'farsReadTest' comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}

