% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Creating a FARS filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{- The year parameter is a numerical variable.}
}
\value{
a filename with the format "accident_<year>.csv.bz2"

#@example
#make_filename(2013)
}
\description{
The 'make_filename' function creates a filename, having a charactor vector containing a formatted combination
of text and the 'year' variable.
}
\details{
When the year is given, the year is converted into a integer by 'as.integer' from the base package, and
assigned as 'year'. Then, the 'year' variable will be used in the C function 'sprintf' from the base package.
}

